refactor: ChatMessageResponse 를 domain/chat/dto/response 로 이동한다.
